[isolated] fix cache race condition
When `isolated download` downloads many duplicated files simultaneously,
it sometimes failed for adding a cache for race condition.

This CL aborts the cache write operation when it detects the cache existence in the mutex section.

Fixed: 1185387
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/2755351
Auto-Submit: Junji Watanabe <jwata@google.com>
Commit-Queue: Junji Watanabe <jwata@google.com>

+++ common/data/caching/cache/cache.go
@@ -352,6 +352,17 @@ func (d *Cache) add(ctx context.Context, digest isolated.HexDigest, src io.Reade
 	d.mu.Lock()
 	defer d.mu.Unlock()
 
+	// If the cache already exists, do not try os.Rename().
+	if d.lru.touch(digest) {
+		logging.Debugf(ctx, "cache already exists. path: %s, digest %s\n", d.path, digest)
+		if cb != nil {
+			if err := cb(); err != nil {
+				return err
+			}
+		}
+		return nil
+	}
+
 	if err := os.Rename(fname, d.itemPath(digest)); err != nil {
 		_ = os.Remove(fname)
 		return errors.Annotate(err, "failed to rename %s -> %s", fname, d.itemPath(digest)).Err()
